+++ python/ray/rllib/agents/impala/vtrace.py
@@ -34,7 +34,7 @@ from __future__ import print_function
 
 import collections
 
-from ray.rllib.models.action_dist import Categorical
+from ray.rllib.models.tf.tf_action_dist import Categorical
 from ray.rllib.utils import try_import_tf
 
 tf = try_import_tf()
