+++ python/ray/rllib/agents/ppo/test/test.py
@@ -6,7 +6,7 @@ import unittest
 import numpy as np
 from numpy.testing import assert_allclose
 
-from ray.rllib.models.action_dist import Categorical
+from ray.rllib.models.tf.tf_action_dist import Categorical
 from ray.rllib.agents.ppo.utils import flatten, concatenate
 from ray.rllib.utils import try_import_tf
