+++ python/ray/rllib/models/__init__.py
@@ -1,9 +1,10 @@
+from ray.rllib.models.action_dist import ActionDistribution
 from ray.rllib.models.catalog import ModelCatalog, MODEL_DEFAULTS
 from ray.rllib.models.model import Model
 from ray.rllib.models.preprocessors import Preprocessor
-from ray.rllib.models.fcnet import FullyConnectedNetwork
 
 __all__ = [
+    "ActionDistribution",
     "ModelCatalog",
     "Model",
     "Preprocessor",
