+++ python/ray/rllib/models/catalog.py
@@ -11,11 +11,11 @@ from ray.tune.registry import RLLIB_MODEL, RLLIB_PREPROCESSOR, \
     _global_registry
 
 from ray.rllib.models.extra_spaces import Simplex
-from ray.rllib.models.action_dist import (Categorical, MultiCategorical,
-                                          Deterministic, DiagGaussian,
-                                          MultiActionDistribution, Dirichlet)
 from ray.rllib.models.torch.torch_action_dist import (TorchCategorical,
-                                                TorchDiagGaussian)
+                                                      TorchDiagGaussian)
+from ray.rllib.models.tf.tf_action_dist import (
+    Categorical, MultiCategorical, Deterministic, DiagGaussian,
+    MultiActionDistribution, Dirichlet)
 from ray.rllib.models.tf.modelv1_compat import make_v1_wrapper
 from ray.rllib.models.preprocessors import get_preprocessor
 from ray.rllib.models.fcnet import FullyConnectedNetwork
